requirements.txt
streamlit==1.38.0
sqlalchemy==2.0.34
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pyjwt==2.9.0
pytesseract==0.3.13
pillow==10.4.0
pandas==2.3.3
requests==2.32.3
numpy==2.1.1
matplotlib==3.9.2
plotly==5.24.1
reportlab==4.2.2
email-validator==2.2.0
httpx==0.27.2
